# news_api_plugin/openapi.yaml
openapi: 3.0.1
info:
  title: Get Statistics Plugin
  description: A plugin that allows the user to get a list of news related to a specific topic. This topic will be chosen by the user. The user will be able to select a title and all the information corresponding to that news will be returned. User will be able to supply an url of an article or pdf in order to get information.
  version: 'v1'
servers:
  - url: http://localhost:8080
paths:

  /select_topic:
    post:
      operationId: selectTopic
      summary: Get a list of titles news related to the topic.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/selectTOPIC'
      responses:
        "200":
          description: OK

  /select_title:
    post:
      operationId: selectTitle
      summary: Select one title listed, then obtain several general infromation of the new.
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/selectTITLE'
      responses:
        "200":
          description: OK

  /get_url_information:
    post:
      operationId: getUrlInformation
      summary: Supply url of an article in order to get information.
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/getURLINFORMATION'
      responses:
        "200":
          description: OK

  /get_pdf_information:
    post:
      operationId: getPDFInformation
      summary: Supply url of a pdf in order to get information.
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/getPDFINFORMATION'
      responses:
        "200":
          description: OK



components:
  schemas:
    selectTOPIC:
      title: topic name
      type: object
      properties:
        topic:
          type: string
          description: topic choosen.
          required: true
        
    selectTITLE:
      title: Title name
      type: object
      properties:
        title:
          type: string
          description: title selected.
          required: true

    getURLINFORMATION:
      title: Url
      type: object
      properties:
        title:
          type: string
          description: article url.
          required: true

    getPDFINFORMATION:
      title: Url
      type: object
      properties:
        title:
          type: string
          description: pdf url.
          required: true